[LEDE] We need to set EXT4_BG_INODE_ZEROED on at least Block Group 0.
This bit says whether the inode table has been explicitly zeroed
or not. ext4 in the kernel now insists that this bit be set on
block group 0 at least.

Since we are creating the filesystem as a flat sparse file, by
definition, the inode tables are zeroed on creation.

* * *

Provenance information, for Mobile NixOS:

This patch comes from the OpenWRT forum.

 * https://forum.openwrt.org/t/read-only-file-system-after-a-fresh-installation-of-lede-17-01-5/17242/8

The patch should come from:

 * `tools/make-ext4fs/patches/200-set-EXT4_BG_INODE_ZEROED.patch`

Though, it seems it never has gone through to LEDE, neither into
OpenWRT's make_ext4fs project...